refactor: align with Node directory structure
