+++ node/_crypto/crypto_browserify/public_encrypt/mgf.js
@@ -1,7 +1,7 @@
 // Copyright 2018-2022 the Deno authors. All rights reserved. MIT license.
 // Copyright 2017 Calvin Metcalf. All rights reserved. MIT license.
 
-import { createHash } from "../../hash.ts";
+import { createHash } from "../../../internal/crypto/hash.ts";
 import { Buffer } from "../../../buffer.ts";
 
 export default function (seed, len) {
