+++ node/_crypto/crypto_browserify/public_encrypt/private_decrypt.js
@@ -2,7 +2,7 @@
 // Copyright 2017 Calvin Metcalf. All rights reserved. MIT license.
 
 import parseKeys from "../parse_asn1/mod.js";
-import { createHash } from "../../hash.ts";
+import { createHash } from "../../../internal/crypto/hash.ts";
 import mgf from "./mgf.js";
 import { xor } from "./xor.js";
 import { BN } from "../bn.js/bn.js";
