--- node/_crypto/types.ts
@@ -1,4 +1,0 @@
-// Copyright 2018-2022 the Deno authors. All rights reserved. MIT license.
-import { Buffer } from "../buffer.ts";
-
-export type HASH_DATA = string | ArrayBufferView | Buffer;
