+++ node/internal/crypto/_randomInt_test.ts
@@ -1,6 +1,6 @@
 // Copyright 2018-2022 the Deno authors. All rights reserved. MIT license.
-import randomInt from "./randomInt.ts";
-import { assert, assertThrows } from "../../testing/asserts.ts";
+import randomInt from "./_randomInt.ts";
+import { assert, assertThrows } from "../../../testing/asserts.ts";
 
 const between = (x: number, min: number, max: number) => x >= min && x < max;
