+++ node/internal/crypto/scrypt_test.ts
@@ -1,7 +1,7 @@
 // Copyright 2018-2022 the Deno authors. All rights reserved. MIT license.
 import { scrypt, scryptSync } from "./scrypt.ts";
-import { Buffer } from "../buffer.ts";
-import { assertEquals } from "../../testing/asserts.ts";
+import { Buffer } from "../../buffer.ts";
+import { assertEquals } from "../../../testing/asserts.ts";
 
 Deno.test("scrypt works correctly", () => {
   scrypt("password", "salt", 32, (err, key) => {
